rename `genomics.py` to `utils.py`
